+++ lib/trailblazer/operation/present.rb
@@ -1,8 +1,19 @@
-module Trailblazer
-  # Backward-compatibility for 1.1.
-  module Operation::Present
-    def present(*args)
-      build_operation(*args)
+class Trailblazer::Operation
+  module Present
+    def self.included(includer)
+      includer.extend Method
+      includer.| Stop, before: Call
+    end
+
+    module Method
+      def present(params={}, options={}, *args)
+        call(params, options.merge("present.stop?" => true), *args)
+      end
     end
   end
+
+  # Stops the pipeline if "present.stop?" is set, which usually happens in Operation::present.
+  Present::Stop = ->(input, options) { options[:skills]["present.stop?"] ? ::Pipetree::Stop : input }
 end
+
+# TODO: another stop for present without the contract!